requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
